refactor: modularize 3.7K-line imageGeneration.ts (Issue #3594)
- Splits the monolithic image generation handler into 13 domain-specific files
- Extracts provider-specific implementations (OpenAI, Gemini, Stability, Fal, BFL, etc.)
- Uses folder-as-module pattern with index.ts facade for zero backward-compatibility breakages
- Auto-prunes and manages all imports for strict typechecking
